requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
langchain
langgraph
langchain-community
langchain-openai
youtube-transcript-api
tiktoken
google-api-python-client